hourly rate to decimal

+++ Sourcecode/API/Strive.BusinessEntities/Model/EmployeeHourlyRate.cs
@@ -20,7 +20,7 @@ namespace Strive.BusinessEntities.Model
         public int? LocationId { get; set; }
 
         [Column]
-        public int HourlyRate{ get; set; }
+        public decimal HourlyRate{ get; set; }
        
         [Column]
         public bool? IsActive { get; set; }
